backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
mysql-connector-python==9.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.40.0
httpx>=0.25.0,<1.0.0
Pillow==10.1.0
pdf2image==1.16.3
aiofiles==23.2.1
jinja2==3.1.2
